👩‍🏭(paths): revert to same paths as Open edX for installation
When we implemented the first Dockerfile POC, it seemed easier
to copy app files in "/app" and data files in "/data" as is
often seen in Dockerfiles. We can avoid a lot of tweaks and
and be friendlier to the Open edX community by sticking to the
paths they use "/edx/app" and "/edx/var".