👷(tags) push docker images with tags matching their git tag
We want to push Docker images to Docker Hub only when a
release is tagged. The git tag is used as the Docker image
tag. We have several lines of images, each one will live
on a branch (e.g. ginkgo.1, ginkgo-funwb,...) and we check
that tags are of the form {branch name}-{semver}.