🔧(lms) derive lms setting files from common.py
We need to have a better control of our settings
than overriding aws which is designed for hosting
on... aws. This commit proposes default settings
copied from the aws.py file and adapted for Docker.
We switch to yaml files for customization instead
of json files because it's cleaner and allows
comments.