⬆️(edxapp) upgrade to hawthorn
A lot of changes were done to 1) the way dependencies
are managed 2) the build of frontend assets 3) settings.
It is not possible anymore to optimize the order in which
dependencies are installed. Assets are not built
differently for development and for production so we
re-run "update_assets" in the development Dockerfile.
Settings are scattered across the code base hardcoding
references to edX's own settings files so we had to use
some tricks and even submit a PR.