🐛(statics) remove webpack stats from STATIC_ROOT
There is no obvious reason to store webpack-stats.json files in the
STATIC_ROOT tree. It causes more problem than it solves as we need to
manually copy it during deployment since it is not collected via
collectstatics. Generating those files with update_assets at the project
root seems a reasonnable choice.