⏪(statics) revert webpack loader settings
In fact, changing django settings is not sufficient to change the
webpack-stats.json path. We also need to update front-end settings that
enforce the usage of the STATIC_ROOT directory as the target path to
generate it.

For now, we will wait for edx to fix it, and we will declare a new issue
to let them know what problem we encountered.