🐛(settings) fix platform name and description default values
In Open edX, the PLATFORM_NAME and PLATFORM_DESCRIPTION settings
are set to lazy translatable strings by default but some code
in Open edX tries to serialize them with a standard json serializer
which breaks. The only way to fix it is to submit a PR to Open edX.
In the mean time we configure these settings with standard strings
as default values.