🔧(databases) improve flexibility of mysql and mogodb settings
We want to be able to configure separately each part of the
database settings (e.g. db name versus password) so that they
can come from different sources and be injected differently.
For example the password comes from a vault and is injected
via environment variables versus the database name comes from
a yaml file and is injected by mounting the file in the container.