⬆️(compose) refactor volume mounts to respect development/production
Static files are not the same in the production build and in the
develoment build. For example production files are processed with
to include a hash of their content in their name for caching. The
webpack manifest file "webpack-stats.json" is therefore also different.
We must therefore provice separate locations to store the static
files of each environment.

This commit also adds missing volume mounts that surface after we
switched to running containers with the local user.