🚸(config) apply the formatter for all sources of settings
We were only applying the formatter on settings set via an
environment variable. Settings coming from the yaml file may also
come with a wrong type so it is more robust to also apply the
formatter on them (for example Mongodb fails if the port is given
as a string instead of an integer). We only apply the formatter
to string values because yaml files can also set settings as lists
or dictionary which we should respect, versus environment variables
who pass the same as a json object...