🐛(compose) fix development entrypoint
When working with development images, the entrypoint (that wraps
commands execution in a python virtual environment), is not properly
called and the resulting command is not executed in the virtual
environment context.

I have the feeling that the commands is spawn in a new shell that does
not inherit from the entrypoint's environment.

Redefining the entrypoint in the docker-compose.yml file seems to fix
the issue.