➕(core) add a requirement file for Open edx Extended
This requirement file was originally located in the oee fork of
edx-platform in openfun/edx-platform. In fact the small fork works
normally if the configurable LTI consumer is not installed. It means
the fork is robust and we can handle dependencies in openedx-docker.
This is easier and makes more sense for all the dependencies that we
need to add that have nothing to do with the edx-platform fork.