✨(project) add environment management tools
To ease working on a particular release, we've defined an activation
script for each release (e.g. an activate script at the root of each
release folder); the aforementioned script updates the working
environment to target this release.

We've also added an utility to assist / ease quick start activation: use
bin/activate to list select release you want to work on.