✨(project) add bin/compose shortcut
When using raw docker-compose commands, the UID and GID environment
variables are supposed to be defined. Hence, to avoid having a warning
message (and potential errors), one should use the following command:

UID=$(id -u) GID=$(id -g) docker-compose

To ease our life, we've added a bin/compose script that can be used as a
shortcut for a configured docker-compose command.