🐛(project) prevent shell built-in vars collisions
Since we use Bash in the Makefile, we cannot set the UID variable (which
is read-only). Previous implementation was raising an error and user
definition in docker-compose was no longer effective.

By explicitly renaming the UID variable to DOCKER_UID, we prevent
such variable collision and the docker user definition is restored.

We've also decided to rename the GID variable to DOCKER_GID for
consistency purpose even if it wasn't required at all.