✨(project) add database cleanup task
When working on different releases, one need to drop/reset databases to
prevent schema inconsistency errors. We've added a clean-db Makefile
task to perform this cleanup. We've also decided to declare this new
task as a bootstrap dependency. Hence, the good way to switch from a
release to another would be:

1. make stop
2. activate a release via: bin/release
3. make bootstrap