🍎(project) improve Bash scripts POSIX compatibility
"sed -r" option is not available on MacOSX, it is recommended to use the
"-E" option to improve POSIX compatibility.

The "realpath" command is part of the GNU/coreutils package that is not
installed by default on MacOSX. We've removed the function call to
something less convoluted to avoid having a core dependency for such a
small achievement.

Fix #117