✨(project) make sure we clean files when bootstrapping
If we don't ensure a full consistency between local sources and the build, we
may broke the development environment with old sources locally fetched
and recent sources in the build.

Thus the bootstrap target is now the way to ensure build reproducibility
when starting to work on a release.

Fix #158