⬆️(eucalyptus/3/*) upgrade to nodejs 10
We recently encountered installation issues with the edx-ui-toolkit
package. Many installed packages are not compatible with the system
nodejs release. Upgrading to a recent nodejs release (as we do for
dogwood releases) seems to fix that issue.

FTR besides deprecated packages the following packages are no longer
compatible with system nodejs release:

- WARN deprecated formidable@1.0.17: Old versions of Formidable are not
  compatible with the current Node.js; Upgrade to 1.2.0 or later
- WARN engine atob@2.1.2: wanted: {"node":">= 4.5.0"} (current:
  {"node":"4.2.6","npm":"3.5.2"})
- WARN engine http-proxy@1.18.0: wanted: {"node":">=6.0.0"} (current:
  {"node":"4.2.6","npm":"3.5.2"})