🔧([dogwood|eucalyptus]/3/[fun|wb]) relocate certificates to media files
These certificates are specific to the dogwood.3/fun image. Saving them
to a specific location forces us to add a volume only this flavor, which
we are reluctant to do because we would like to have a unique edxapp
application in github.com/openfun/arnold able to deploy any flavor.

We noticed that certificates are in fact media files (they are generated
dynamically, pointed by a db record and served by nginx), so saving them
to a specific folder within media files is simpler and makes sense.