♻️(flavors/*) refactor way to configure authentication backends
Open edX is trying to ease authentication backends configuration but
it is making it convoluted and difficult to figure out what we
actually do. We simplify authentication backend configuration by
allowing projects to define the setting either completely or not at
all. In this last case, we fall back on the default backend.