♻️(dogwood/3/fun) use manifest static files storage
We recently moved static files to the "nginx-edxapp" image but, because
edX is using the CachedStaticFilesStorage, we still had to make staticfiles
available to edxapp by mounting the staticfiles volume and running the
collectstatic command upon each deployment. This can be avoided by using
a storage backend based on a static files manifest.