♻️(flavors/*) include static files in edxapp image
We recently introduced a custom edxapp-nginx image that includes all
static files to make serving static files more efficient (direct from
the k8s node instead of via a PVC). But edxapp still needed to have
access to all static files in order to calculate the files' hash and
compute their url. Since the static files were not collected inside
the edxapp image, we had to mount a staticfiles pvc and run
collectstatic upon each deployment... very slow!

This PR runs collectstatic in the edxapp image so that all static
files are available at run time. We didn't want the image size to
increase significantly, so the collectstatic is run with the --link
option, so the files are not copied but symlinked. The manifest
version of the files are still symlinked as well by running rdfind.