✨(dockercompose) add bridge network to connect with outside project
We want to connect external projects to Open edX via SSO. One of the requests
is made server-to-server which requires that both the remote app and the edx
containers to be on the same network.

We connect directly to the nginx container so that static files are available
as well and give it a custom name on the outside network so that its name is
explicit on the remote aop that needs to connect to it.