✨(hawthorn/1/*) add MKTG_URLS setting
We want to use MKTG_URLS. This setting is used by edx when
FEATURES.ENABLE_MKTG_SITE is enabled. MKTG_URLS is a mapping url in json
format, the jey are the same than MKTG_URL_LINK_MAP setting and the
value can be relative or absolute urls. This setting is used by mako
templating in the function marketing_link.