🚚(master-bare) move master-bare to master.0-bare
Move master/bare to master/0/bare structure to be in line
with other release naming format.

This is done to enable the circle-ci pack feature. The script
that prepare the fragments for each release is scraping path
in the release folder. Without this change we get a
`master/bare/src` name instead of `feature.version-flavor`
naming.