🔨(project) prevent make bootstrap to be run as root user
In order to not mess up user filesystem, the build should not be
run as a root user.
This commit add a "check-root-user" make target that will check
if the user running the make is not with an id = 0 (root).
For future use, this new target can be added to any
make target if needed.