🐛(ironwood/2) fix an env settings that was left aside
In the docker_run_production.py, the LMS_BASE and CMS_BASE point
to the port 8072 and 8082. That's ok for dev but not compliant
with prod settings. To solve this, an env.d/production file is
created and injected in the docker-compose config to override
this setting in a clean manner.