♻️(ci) stop pulling images anonymously
Docker will start a rate limit on pulling images on docker hub. This rate
limit is based on the IP used for anonymous users, this is the CircleCI IP
that will be used so the rate limit will be reached quickly. If the image pull
is made with a connected user, the rate limit becomes a per-user rate so we
should not reach it once connected.