🔧(Makefile) add check-activate task
By default activated OpenEdx release was master/bare. But launch commands
without the right activated release happend too often and it was time consuming.
To address this there is no more default activated OpenEdx release and a new
task check-activate is in charge to check before each command that a release
is activated.