🐛(dogwood/3/fun) fix Docker build
Building the dogwood/3/fun Docker fails with the following error:

    Step 51/51 : RUN NO_PREREQ_INSTALL=1     paver update_assets
      --settings=fun.docker_build_production --skip-collect
    ...
      File "/.../django/contrib/auth/decorators.py",
      line 6, in <module>
        from django.shortcuts import resolve_url
    ImportError: cannot import name resolve_url

The root cause is unknown; what happens is that django==1.8.12 is installed on
top of django==1.4.22 and the shortcuts.py module finds itself right next to
the (obsolete) shortcuts/__init__.py module. The latter does not contain the
`resolve_url` function and causes the failure.

To resolve (pun intended) this, we explicitly uninstall django=1.4.22. This
*should not* cause any new issue, because django==1.8.12 is already the version
currently installed on FUN.

Close #270.