🐛(Dockerfile) update config file even starting from existing image
We recently refactored building dogwood images to start from an
existing edxapp image instead of the ubuntu image from scratch,
because some Python packages could not be found anymore.

In this refactoring we forgot to update the config files so the new
VIDEOFRONT_CDN_BASE_URL setting that was added recently was not
found in the built docker image.