♻️(dogwood/3/fun) revert to building the image from scratch
Since pypi banned clients that don't support SNI, we did not manage
to build the image from scratch and tried to build each image on
top of the previous image.

This strategy causes our images to inflate and is not sustainable.

We now propose to revert building the image from a plain  ubuntu 12.04
image and upgrade its Python 2 to the last version available: 2.7.18.