🔧(Makefile) fix variable declarations
On github the organization edx has been renamed into openedx. As we are
fetching some edx releases from this organization on github, we got 404
response. So we update the url to fix that.

Furthermore, it appears that according to the active release,
environment variables EDX_ARCHIVE_URL and EDX_DEMO_ARCHIVE_URL were set
or not. So if in the same terminal session we switched of release, those
 variables was not unset that leads to misconfiguration.